requirements.txt
speech\Scripts\activate ---- activate the env
deactivate ---- to deactive the env



Flask== 3.0.2
pytesseract==0.3.10
Pillow==10.2.0
googletrans==4.0.0rc1
gTTS==2.5.1
google-generativeai==0.4.1
